Move Variant to ~/common
